+++ PDM_Microphone/PDM_Microphone.cywrk
@@ -2,17 +2,18 @@
 <CyXmlSerializer>
 <!--This file is machine generated and read. It is not intended to be edited by hand.-->
 <!--Due to this, there is no schema for this file.-->
-<CyGuid_2867d519-54d2-4c01-9830-c51cb08bc3dd type_name="CyDesigner.Common.ProjMgmt.Model.CyPrjMgmtWrkspc" version="5" xml_contents_version="2" name="PDM_Microphone" persistent="" platform="NOT_USED">
-<current_project name=".\PDM_Microphone.cydsn\PDM_Microphone.cyprj" />
+<CyGuid_2867d519-54d2-4c01-9830-c51cb08bc3dd type_name="CyDesigner.Common.ProjMgmt.Model.CyPrjMgmtWrkspc" version="5" xml_contents_version="3" name="PDM_Microphone" persistent="" platform="NOT_USED">
+<current_project name=".\PDM_Microphone_P4BLE.cydsn\PDM_Microphone_P4BLE.cyprj" />
 <wrkspc_folders />
 <files />
 <projects>
-<name_val_pair name="PDM_Microphone" v=".\PDM_Microphone.cydsn\PDM_Microphone.cyprj" />
+<v>.\PDM_Microphone_P5LP.cydsn\PDM_Microphone_P5LP.cyprj</v>
+<v>.\PDM_Microphone_P4BLE.cydsn\PDM_Microphone_P4BLE.cyprj</v>
 </projects>
 <workspace_id v="fa6e91b3-680a-463a-a8b5-f33b20c6e424" />
-<WriteAppVersionLastSavedWith v="3.0.0.1539" />
-<WriteAppMarketingVersionLastSavedWith v="3.0" />
-<CyGuid_dcbd9771-0334-43dc-9cc3-fe99dc3c5316 type_name="CyDesigner.Common.ProjMgmt.Model.CyPrjDependencyList" version="3">
+<WriteAppVersionLastSavedWith v="3.2.0.724" />
+<WriteAppMarketingVersionLastSavedWith v="3.2" />
+<CyGuid_dcbd9771-0334-43dc-9cc3-fe99dc3c5316 type_name="CyDesigner.Common.ProjMgmt.Model.CyPrjDependencyList" version="4">
 <wrkspc_dependencies />
 </CyGuid_dcbd9771-0334-43dc-9cc3-fe99dc3c5316>
 <CyGuid_63b68103-67f5-4406-8da6-5c8625765b82 type_name="CyDesigner.Common.ProjMgmt.Model.CyIgnoredSystemDepsList" version="1">
